pyproject.toml
[project]
name = "flyagent"
version = "0.1.0"
description = "Dynamic ICTM-based research agent inspired by AOrchestra — creates specialized SubAgents on the fly"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "html2text>=2024.2",
    "ddgs>=9.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "tomli>=2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]

[project.scripts]
flyagent = "main:main_cli"
